🐛 wait for sdk to load
